git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@3501 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinx.h
@@ -1780,8 +1780,8 @@ struct CSphSourceParams_SQL
 	CSphString						m_sQuery;
 	CSphString						m_sQueryRange;
 	CSphString						m_sQueryKilllist;
-	int								m_iRangeStep;
-	int								m_iRefRangeStep;
+	int64_t							m_iRangeStep;
+	int64_t							m_iRefRangeStep;
 	bool							m_bPrintQueries;
 
 	CSphVector<CSphString>			m_dQueryPre;
